pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "librosa>=0.11.0",
    "matplotlib>=3.10.1",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "psutil>=7.0.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "scipy>=1.15.2",
    "soundfile>=0.13.1",
    "streamlit>=1.44.0",
]
